front: navigate to zone preferences on first login (#788)
