+++ public/mobile-app/src/routes/welcome/zones/page.svelte.test.ts
@@ -1,0 +1,38 @@
+import { fireEvent, render, screen, waitFor } from '@testing-library/svelte';
+import { describe, expect, test, vi } from 'vitest';
+import * as navigationMethods from '$app/navigation';
+import { userStore } from '$lib/state/User.svelte';
+import { mockUserInfo } from '$tests/utils';
+import Page from './+page.svelte';
+
+describe('/+page.svelte', () => {
+  test('user has to be connected', async () => {
+    // Given
+    const spy = vi.spyOn(navigationMethods, 'goto').mockResolvedValue();
+
+    // When
+    render(Page);
+
+    // Then
+    await waitFor(() => {
+      expect(spy).toHaveBeenCalledTimes(1);
+      expect(spy).toHaveBeenCalledWith('/');
+    });
+  });
+
+  test('should navigate homes when user clicks on Passer button', async () => {
+    // Given
+    await userStore.login(mockUserInfo);
+    const spy = vi.spyOn(navigationMethods, 'goto').mockResolvedValue();
+    render(Page);
+
+    // When
+    const button = screen.getByTestId('skip-button');
+    await fireEvent.click(button);
+
+    // Then
+    await waitFor(() => {
+      expect(spy).toHaveBeenCalledWith('/');
+    });
+  });
+});
